+++ app/src/androidTest/java/woowacourse/payments/ui/component/CardExpirationDateTextFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.component
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -13,7 +13,7 @@ import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.payments.CardExpirationDateTextField
+import woowacourse.payments.ui.payments.component.CardExpirationDateTextField
 
 class CardExpirationDateTextFieldTest {
     @get:Rule
