+++ app/src/androidTest/java/woowacourse/payments/ui/component/CardNumberTextFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.component
 
 import android.util.Log
 import androidx.compose.runtime.getValue
@@ -13,7 +13,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.payments.CardNumberTextField
+import woowacourse.payments.ui.payments.component.CardNumberTextField
 
 class CardNumberTextFieldTest {
     @get:Rule
