+++ app/src/androidTest/java/woowacourse/payments/ui/component/CardholderNameTextFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.component
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -14,7 +14,7 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.payments.CardholderNameTextField
+import woowacourse.payments.ui.payments.component.CardholderNameTextField
 
 class CardholderNameTextFieldTest {
     @get:Rule
@@ -72,7 +72,7 @@ class CardholderNameTextFieldTest {
 
         // then
         composeTestRule
-            .onNodeWithText("5/30")
+            .onNodeWithText("ABCDE")
             .assertIsDisplayed()
     }
