+++ app/src/main/java/woowacourse/payments/ui/payments/CardRegistrationActivity.kt
@@ -1,10 +1,10 @@
-package woowacourse.payments
+package woowacourse.payments.ui.payments
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import woowacourse.payments.ui.payments.CardRegistrationScreen
+import woowacourse.payments.ui.payments.screen.CardRegistrationScreen
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 class CardRegistrationActivity : ComponentActivity() {
