+++ app/src/main/java/woowacourse/payments/ui/payments/component/CardRegistrationTopAppBar.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.payments
+package woowacourse.payments.ui.payments.component
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
