+++ app/src/main/java/woowacourse/payments/ui/payments/component/CardholderNameTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.payments
+package woowacourse.payments.ui.payments.component
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
